refactor(cli): Extract duplicate utility functions into shared module
- Create new utils.rs module with shared network utility functions
- Extract is_port_in_use() function to eliminate duplication across check.rs, doctor.rs, and status.rs
- Extract parse_redis_url() function for consistent Redis URL parsing
- Extract is_redis_accessible() function for async Redis connectivity testing
- Update all command modules to use shared utilities from crate::utils
- Remove 49+ lines of duplicate code across 3 files
- Improve maintainability by centralizing common network operations
- Fix async/sync consistency for port checking operations
