+++ crates/cli/src/commands/doctor.rs
@@ -3,6 +3,7 @@ use std::collections::HashMap;
 use std::fs;
 use std::path::Path;
 use std::process::Command;
+use crate::utils::{is_port_in_use, parse_redis_url, is_redis_accessible};
 
 pub async fn run(fix_issues: bool, verbose: bool) -> Result<(), ElifError> {
     println!("🩺 Running elif.rs project diagnostics...");
@@ -416,7 +417,7 @@ impl Doctor {
         }
 
         // Check for essential services if they should be running
-        if std::env::var("DATABASE_URL").is_ok() && !self.is_port_in_use(5432) {
+        if std::env::var("DATABASE_URL").is_ok() && !is_port_in_use(5432) {
             self.issues.push(Issue {
                 category: IssueCategory::FrameworkHealth,
                 severity: IssueSeverity::Critical,
@@ -427,9 +428,9 @@ impl Doctor {
         }
 
         if let Ok(redis_url) = std::env::var("REDIS_URL") {
-            let (host, port) = self.parse_redis_url(&redis_url).unwrap_or(("127.0.0.1".to_string(), 6379));
+            let (host, port) = parse_redis_url(&redis_url).unwrap_or(("127.0.0.1".to_string(), 6379));
             
-            if !self.is_redis_accessible(&host, port).await {
+            if !is_redis_accessible(&host, port).await {
                 self.issues.push(Issue {
                     category: IssueCategory::FrameworkHealth,
                     severity: IssueSeverity::Warning,
@@ -442,7 +443,7 @@ impl Doctor {
 
         // Check for application health
         let common_ports = [3000, 8000, 8080];
-        let app_running = common_ports.iter().any(|port| self.is_port_in_use(*port));
+        let app_running = common_ports.iter().any(|port| is_port_in_use(*port));
         
         if !app_running {
             self.issues.push(Issue {
@@ -457,58 +458,6 @@ impl Doctor {
         Ok(())
     }
 
-    fn is_port_in_use(&self, port: u16) -> bool {
-        use std::net::{TcpListener, SocketAddr};
-        
-        let addr = SocketAddr::from(([127, 0, 0, 1], port));
-        TcpListener::bind(addr).is_err()
-    }
-
-    fn parse_redis_url(&self, redis_url: &str) -> Option<(String, u16)> {
-        // Handle common Redis URL formats:
-        // redis://localhost:6379
-        // redis://user:pass@localhost:6379
-        // redis://localhost:6379/0
-        // localhost:6379
-        
-        if let Ok(parsed_url) = url::Url::parse(redis_url) {
-            // Parsed as a full URL
-            let host = parsed_url.host_str().unwrap_or("127.0.0.1").to_string();
-            let port = parsed_url.port().unwrap_or(6379);
-            Some((host, port))
-        } else if redis_url.contains(':') {
-            // Try to parse as "host:port"
-            let parts: Vec<&str> = redis_url.splitn(2, ':').collect();
-            if parts.len() == 2 {
-                if let Ok(port) = parts[1].parse::<u16>() {
-                    return Some((parts[0].to_string(), port));
-                }
-            }
-            None
-        } else {
-            // Just a hostname, use default port
-            Some((redis_url.to_string(), 6379))
-        }
-    }
-
-    async fn is_redis_accessible(&self, host: &str, port: u16) -> bool {
-        use std::net::TcpStream;
-        use std::time::Duration;
-        
-        // Try to establish a TCP connection to the Redis server
-        let addr = match format!("{}:{}", host, port).parse::<std::net::SocketAddr>() {
-            Ok(addr) => addr,
-            Err(_) => return false,
-        };
-        
-        // Use a short timeout for the connection attempt
-        match tokio::time::timeout(Duration::from_millis(1000), async move {
-            TcpStream::connect(addr)
-        }).await {
-            Ok(Ok(_)) => true,
-            _ => false,
-        }
-    }
 
     fn check_elif_dependencies(&self, parsed_cargo: &toml::Value) -> bool {
         if let Some(deps) = parsed_cargo.get("dependencies").and_then(|d| d.as_table()) {
